third_party/WebKit/LayoutTests/platform/mac/fast/dom/Range/getBoundingClientRect-caret-expected.txt

hello
Test getBoundingClientRect of caret for collapsed range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL rect.left should be 22. Was 23.
PASS rect.top is 8
PASS rect.width is 0
FAIL rect.height should be 19. Was 18.
PASS successfullyParsed is true

TEST COMPLETE
